LayoutTests/fast/files/xhr-response-blob-expected.txt

Test that XHR.responseType = 'blob' gives you back a Blob.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.responseType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
PASS xhr.response is null
PASS xhr.response is null
PASS xhr.response instanceof Blob is true
PASS xhr.response.type is "text/plain"
PASS xhr.responseType is "blob"
PASS xhr.response is null
PASS xhr.responseType is "blob"
PASS xhr.response is null
PASS xhr.response instanceof Blob is true
PASS xhr.response.type is ""
